# swagger.yaml
swagger: "2.0"
info:
  title: Olibo League API
  description: |
    ## API de gestion de la Ligue Olibo

    Cette API permet de gérer intégralement une compétition de football : équipes, matchs,
    licences, votes, paiements, médias et bien plus.

    ### Authentification
    La plupart des endpoints nécessitent un **JWT Bearer Token**.
    Connectez-vous via `POST /api/auth/login` pour obtenir vos tokens,
    puis cliquez sur **Authorize** (🔒) et entrez : `Bearer <votre_access_token>`

    ### Rôles disponibles
    | Rôle | Description |
    |------|-------------|
    | `super_admin` | Accès total |
    | `admin_competition` | Gestion des compétitions et équipes |
    | `operator` | Gestion des matchs |
    | `referee` | Arbitre — feuilles de match et incidents |
    | `commissioner` | Commissaire de match |
    | `team_captain` | Capitaine d'équipe |
    | `coach` | Entraîneur |
    | `player` | Joueur |
    | `spectator` | Spectateur |
    | `journalist` | Journaliste |

  version: "1.0.0"
  contact:
    name: Équipe Olibo
    email: contact@olibo.com
  license:
    name: Propriétaire
host: "127.0.0.1:8000"
basePath: "/"
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Format requis : **Bearer {access_token}**"
...
definitions:
...
  License:
    type: object
    properties:
      id:
        type: integer
        example: 1
      member_id:
        type: integer
        example: 10
      season_id:
        type: integer
        description: Saison à laquelle la licence est rattachée
        example: 1
      license_number:
        type: string
        description: "Généré automatiquement au format OL-{season.label}-{team_id:03d}-{member_id:03d}"
        example: OL-2025/26-001-010
      issue_date:
        type: string
        format: date
        example: "2025-08-01"
      expiry_date:
        type: string
        format: date
        example: "2026-06-30"
      is_valid:
        type: boolean
        description: La licence est documentairement valide
        example: true
      is_active:
        type: boolean
        description: La licence est active pour la saison en cours (false si saison révoquée)
        example: true
      document_url:
        type: string
        example: https://example.com/docs/license.pdf
      created_at:
        type: string
        format: date-time

  # --- Match ---
...
  Error:
    type: object
    properties:
      error:
        type: string
        example: Unauthorized
      message:
        type: string
        example: Token JWT invalide ou expiré

# ============================================================
# PATHS
# ============================================================
paths:
...
  /api/license:
    post:
      tags: [License]
      summary: Créer une licence
      description: |
        Crée une licence pour un membre de type `player` uniquement.
        Réservé aux `super_admin` et `admin_competition`.
        Le numéro de licence est **auto-généré** au format `OL-{season.label}-{team_id:03d}-{member_id:03d}`.
        Il est possible de préciser un `season_id` ; sinon la saison active est utilisée.
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [member_id, issue_date, expiry_date]
            properties:
              member_id:
                type: integer
                example: 10
              season_id:
                type: integer
                description: Saison concernée (optionnel — saison active par défaut)
                example: 1
              issue_date:
                type: string
                format: date
                example: "2025-08-01"
              expiry_date:
                type: string
                format: date
                example: "2026-06-30"
              document_url:
                type: string
                example: https://example.com/license.pdf
      responses:
        201:
          description: Licence créée
          schema:
            $ref: '#/definitions/License'
        400:
          description: Membre non joueur ou licence déjà existante
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Permissions insuffisantes
    get:
      tags: [License]
      summary: Lister les licences
      description: |
        Filtrées selon le rôle :
        - Admin : toutes les licences
        - Capitaine/Coach : licences de son équipe uniquement
      security:
        - Bearer: []
      parameters:
        - in: query
          name: member_id
          type: integer
          description: Filtrer par membre
        - in: query
          name: team_id
          type: integer
          description: Filtrer par équipe
        - in: query
          name: season_id
          type: integer
          description: Filtrer par saison
        - in: query
          name: is_valid
          type: boolean
          description: Filtrer par validité documentaire
        - in: query
          name: is_active
          type: boolean
          description: Filtrer par statut actif (false = saison révoquée)
      responses:
        200:
          description: Liste des licences
          schema:
            type: object
            properties:
              total:
                type: integer
              licenses:
                type: array
                items:
                  $ref: '#/definitions/License'
  /api/license/season/{season_id}/team/{team_id}:
    get:
      tags: [License]
      summary: Licences d'une équipe pour une saison donnée
      description: |
        Retourne les licences actives de l'équipe pour la saison précisée.
        Le header `X-Active-Season` indique le label de la saison actuellement active.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: season_id
          required: true
          type: integer
        - in: path
          name: team_id
          required: true
          type: integer
      responses:
        200:
          description: Liste des licences de l'équipe pour cette saison
          headers:
            X-Active-Season:
              type: string
              description: "Label de la saison active (ex: 2025/26)"
          schema:
            type: object
            properties:
              total:
                type: integer
              licenses:
                type: array
                items:
                  $ref: '#/definitions/License'
        404:
          description: Saison introuvable
  /api/license/team/{team_id}/generate:
    post:
      tags: [License]
      summary: Générer les licences de tous les joueurs d'une équipe
      description: |
        Crée automatiquement une licence pour chaque joueur actif de l'équipe
        qui n'en possède pas encore pour la saison cible.
        Si aucun `season_id` n'est fourni, la saison active est utilisée.
        Les dates `issue_date` / `expiry_date` sont déduites des dates de la saison
        si elles ne sont pas fournies explicitement.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: team_id
          required: true
          type: integer
        - in: body
          name: body
          schema:
            type: object
            properties:
              season_id:
                type: integer
                description: "Saison cible (défaut : saison active)"
              issue_date:
                type: string
                format: date-time
                description: "Date d'émission (défaut : début de saison)"
              expiry_date:
                type: string
                format: date-time
                description: "Date d'expiration (défaut : fin de saison)"
      responses:
        201:
          description: Licences générées
          schema:
            type: object
            properties:
              message:
                type: string
                example: "3 license(s) generated"
              created:
                type: array
                items:
                  $ref: '#/definitions/License'
              skipped:
                type: array
                items:
                  type: object
                  properties:
                    member_id:
                      type: integer
                    reason:
                      type: string
        400:
          description: Dates invalides
        403:
          description: Unauthorized
        404:
          description: Équipe ou saison introuvable
  /api/license/{id}:
    get:
      tags: [License]
      summary: Détails d'une licence
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: Données de la licence
          schema:
            $ref: '#/definitions/License'
        404:
          description: Licence non trouvée
    put:
      tags: [License]
      summary: Renouveler ou modifier une licence
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
        - in: body
          name: body
          schema:
            type: object
            properties:
              expiry_date:
                type: string
                format: date
                example: "2027-12-31"
              document_url:
                type: string
                example: https://example.com/new_license.pdf
      responses:
        200:
          description: Licence mise à jour
          schema:
            $ref: '#/definitions/License'
        403:
          description: Permissions insuffisantes
    delete:
      tags: [License]
      summary: Supprimer une licence
      description: Réservé au `super_admin`.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        204:
          description: Licence supprimée
        403:
          description: Permissions insuffisantes
  /api/license/{id}/validate:
    post:
      tags: [License]
      summary: Valider une licence
      description: Marque la licence comme valide (`is_valid = true`).
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: Licence validée
          schema:
            $ref: '#/definitions/License'
        403:
          description: Permissions insuffisantes
  /api/license/{id}/invalidate:
    post:
      tags: [License]
      summary: Révoquer une licence
      description: Marque la licence comme invalide (`is_valid = false`).
      security:
        - Bearer: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: Licence révoquée
          schema:
            $ref: '#/definitions/License'
        403:
          description: Permissions insuffisantes

  # ==================== MATCH ====================
...